backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
openai==1.3.0
httpx==0.25.1
pydantic==2.5.0
retell-sdk==4.48.0
